refactored dashboard settings into its own component
